Hide max-processing-minutes and improve language
This commit improves the language some for the generated help text for
max-processing-minutes by disambiguating fallback from upstream. This
also hides the flag. This flag is used for async migrations which we
aren't encouraging users to use right now, so it's probably better to
just hide this flag to create an easier user experience.

Change-Id: I9d60fbfa4c1d614fc1931e8fb8f5b9f0e88d2953
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3675
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>