replication: Log when replication slots are created/dropped
This adds additional logging around creating/dropping replication
slots, as the title says.

Additionally, this adjusts the drop function to convert "errors"
relating to the slot not existing in to Ok(()) values. This lets
calling code handle the remaining errors, instead of ignoring them
all like before. Having to compare strings to implement this is
non-ideal, but even refactoring to propagate up the original
Postgres error would still require some fiddling to perform the
check.

Refs: ENG-2075
Change-Id: Ife792b09a28f3f31f7d1dc0361ebcbd2c94806ce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3685
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>