dataflow-expression: Pass types into binary op from_sql_op function
These aren't used for anything yet, but there are cases where we will
want to infer different Dataflow operators for the same type of AST node
by looking at the types of the arguments.

The types passed in via the calls in the noria_connector code should be
correct, though it doesn't really matter what they are right now since
none of the relevant binary operators are likely to overlap with the
ones used in that code. That said, still better to pass in the right
type just in case it matters in the future.

Change-Id: Ied214bfdfeb4dd879b4bd4dd59d5909694caaa8f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3690
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>