replicators: notify after catch up for postgres
If a snapshot completes some but not all tables, fails, then retries at
a higher replication offset and succeeds, different tables can be
at different replication offests. The tables that initially succeeded
will be at a lower replication offset than the ones that succeeded on
the retry. To account for this, we do a catch up action at the end
of snapshotting to synchronize them.

This change notifies waiters after this catch up process for postgres,
to align it with mysql--though tables being at different offsets may
technically be within our consistency model currently, it's better to
have them synchronized if possible.

Change-Id: I7d79164fa7ee899de99b00dde2975be87b92fb78
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3691
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>