dataflow-expression: Add JsonSubtract operator
This new operator represents the `-` operator as applied to JSONB. As
such, we now branch on the left operand's type during lowering time to
determine whether to convert the `Subtract` AST node to the existing
`Subtract` Dataflow operator or the new `JsonSubtract` Dataflow
operator.

For now, nothing is implemented to evaluate this operator; that will be
handled in a future commit.

Change-Id: Ib4327554af43f0edcff5560e0e066b3cac7f2f7d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3692
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>