server: Make the fatal replicator error a little louder
If we get some sort of unhandled error in replication, the replicator
will restart after a timeout - this is great for external transient
errors, but in the case that we've accidentally left a *permanent*
error (such as failing to parse some statement) unhandled (eg ENG-2106),
this means that the only thing we can do is restart the replicators
indefinitely.  This error was getting logged, but the log message was a
little unobtrusive so was easy to get lost - this makes it a little bit
louder, and targets it at the "replicators" crate so that filters for
replicator logs catch it too.

Change-Id: If82fbe585869738e973dc12ad8ae4a0834590bc4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3697
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI