eval: Fix formatting issues
The presence of the comment in the middle of the pattern was triggering
https://github.com/rust-lang/rustfmt/issues/4119, which meant that none
of the code in that match statement was being formatted by rustfmt. This
commit changes the existing code to avoid triggering that bug, and also
reformats the code in that block now that formatting is no longer
silently failing.

Change-Id: Idef7520a18a054f75adbab857823154c37fb270e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3700
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>