dataflow-expression: Add more bidirectional indexing utility functions
The `index_to_bidirectional` function simplifies index operations in
this module.

The `index_bidirectional_mut` function is simply a mutable counterpart
to `index_bidirectional`.

The `remove_bidirectional` function removes an element from a `Vec`.
Unlike `Vec::remove`, it returns an optional value to avoid panicking.

These help implement operations that allow for mutably indexing from the
end of an array with a negative integer, such as with the JSONB
operators `-` and `#-` in PostgreSQL.

Change-Id: I4af0bd65e5cbace90fb466d9ad13bb03a04f6b44
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3727
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>