eval: Implement JsonSubtract for integer RHS
Passing an integer as the right-hand operand to JSONB `-` should remove
the element at that index of a JSON array on the left-hand side.
Negative indexes should offset backward from the end of the array, and
out-of-bound indexes should be ignored, with the same JSON value being
returned as-is.

Implementations for other right-hand side operand types are still to
come in future commits.

Change-Id: I45fa347855ce079130f18a1aafc636a7bd4812b5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3729
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>