eval: Implement JSON `-` for string arrays from JSON arrays
This implements the case where an array of strings is subtracted from a
JSONB value that contains a JSON array. In this case all strings in the
string array should be removed from the JSON array.

Note that we're doing this in O(N*M) time since we remove each key from
the array one at a time. This is probably not a huge deal if we assume
that users normally subtract only small numbers of keys at a time, but
this could be optimized in the future if/when it becomes a bottleneck
for a customer.

Change-Id: Id2375e7434566c6b2342a7cf4f893caa100b1354
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3741
Tested-by: Buildkite CI
Reviewed-by: Nikolai Vazquez <nikolai@readyset.io>