replicators: Schema-qualify enum names in snapshot
Since we don't pass a schema search path when sending CREATE TABLE
statements to extend_recipe, we need to schema-qualify the names of enum
types in those statements (otherwise they won't resolve to the actual
type definition on creation). This does that, clarifying some of the
variable names along the way and also adding a test that tables with
enum columns get properly snapshot.

Change-Id: I9143e1a82ad3d058c9a0f5aa2dcb431ba88b4fe7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3732
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>