sql-passes: Include aliased joined-to subqueries in tables
Include the aliases for joined subqueries in the list of tables that're
being selected FROM in the implied-tables rewrite pass, so that columns
explicitly projected from those subqueries get properly qualified with
the alias for the subquery they come from. This fixes the query in the
docs example (!!), which is also included as a (data-free) logictest.

Fixes: ENG-2070
Change-Id: I894334b2a66ae345ef2cdd1901cc3e3b3a834529
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3759
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>