sql-passes: Consider alias when resolving tables for cols
In the implied_tables pass, when we store the list of tables that are
currently referenced in the query, we currently store the *aliases* for
those tables - but when trying to resolve the tables in the schema, we
only resolve their proper names. This causes cause un-qualified columns
from aliased tables to not resolve. This fixes that, by instead storing
a mapping from the original name of the table to the alias for that
table, using the alias when rewriting the column to be qualified.

Also, this adds docstrings for all the fields on
ExpandImpliedTablesVisitor, which was long overdue.

Release-Note-Core: Fixed a bug where unqualified column references in
  queries wouldn't automatically be qualified if the table they belong
  to had an alias.
Fixes: ENG-1990
Change-Id: I77bb38821c0e703ab5fe9df20f2e2ecf9d1f45a4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3760
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>