mir: Simplify `column_specs` field in `MirNodeInner::Base`.
The `column_specs` field in `MirNodeInner::Base` was a vector of tuples
`(ColumnSpecification, Option<usize>)`, with the second element of the
tuple representing "the absolute column index".
That was probably meant to be used in a context where base tables could
be modified, which we currently don't support.
So, this commit gets rid of the tuple and instead simplifies the field
to be a `Vec<ColumNSpecification>`, and we use the index of the columns
inside that vector instead of the old tuple's second field.

Change-Id: I24f5562009f9c8c683e742303f5e157b9c7a251a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3765
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>