replicators: Fix snapshotting error caused by wrong column order
While the query executed by TableEntry::get_columns often happens to
return its results in the same order as the actual table columns, this
is not guaranteed, and I have found a case in which it returns them in a
different order.

In the example I found, this caused replication to fail due to an
"invalid buffer size" error. After much debugging, I found that this was
because we were passing in a buffer with string data and trying to
convert it to an INT4, which was ultimately caused by us using different
orders for the columns vs the actual data.

To fix this, this commit simply adds an ORDER BY clause to the query
that we use to get the column info.

Change-Id: I114a22bb66c8bcd4506988340e79442841ef8818
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3778
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI