replicators: Allow TableFilter to deny tables
This extends TableFilter to allow for it to also support denying
replication for a set of filters. This will not be exposed as a CLI like
`--replication-tables` is, but will rather be used internally to ignore
tables that we fail to replicate or snapshot, so that the replication
task doesn't continuously crash. Failure to snapshot or replicate a
table is bad, and we should fix it--but it's better to keep working
without replicating those tables than get into a crash loop trying and
failing to replicate or snapshot.

This commit also removes the responsibility of keeping a list of tables
to replicate from TableFilter. TableFilter now just tells you if a table
should be replicated or not.

The mysql snapshot code, which was using the table filter to store the
list of tables instead separately gets the list of tables and filters
them, similar to how the postgres snapshot code was already behaving.

Change-Id: I170068f1af6ee083376bb04b688681e3f3b70117
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3474
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>