psql-srv: Don't decode enums as indices
Internally, we represent enum values for both mysql and postgresql as
the 1-based index of the variant within the list of variants. To match
this, we had previously been parsing enum values *at the protocol level*
as those indices - this works fine for parameters for queries that're
executed against ReadySet, but breaks when we try to then use that
parameter to execute a query that's proxied upstream.

Instead, this commit just leaves the enum as its pg-native text
representation, - when we actually execute a ReadySet query with this
we'll coerce it anyway, so this keeps readyset working but makes enum
parameters for proxied queries work.

Release-Note-Core: Fixed a bug that would cause a protocol error if an
  unsupported query with a parameter with a custom enum type was proxied
  upstream in Postgres
Fixes: ENG-2122
Change-Id: I42d99ac160aab6cc01663cd5eb6fe4807d701ea9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3796
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>