nom: Allow error handling for ALTER TABLE parsing
At this moment, we don't fully support parsing all ALTER TABLE variants,
 only a few of them. This is a problem, because for replication, we
 first parse the statements we get from upstream, and then decide that
 if the statement is an ALTER TABLE, we should trigger a resnapshot (we
 don't even care about the statement itself!).
So, not being able to parse all variants resulted in the replicators
skipping `ALTER TABLE` statements instead of triggering a resnapshot
like they should.

This commit modifies the `AlterTableStatement` struct to have a
result as the `definitions` field, instead of the old
`Vec<AlterTableDefinition>`.
If we were able to parse, that `Result` will be an `Ok` with said
vector.
If we failed to parse, that `Result` will be an `Err` with a `String`
representing the slice that couldn't be parsed.

Refs: ENG-2083
Change-Id: I1e3e6f66eab1b33e2264f1f0585956ed62c81483
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3806
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>