Inlining: Create ReadySet PrepareResult variant without schema
Splits the PrepareResult::Select variant into one with a schema and one
with only the statement id. This allows us to return a prepare result
for a statement that reuses the view of another cached query, as we
cannot determine the schema in this case.

The new NoSchema variant will be returned only if we are able to get an
Upstream result for the prepared statement. Otherwise, we would not be
able to return a valid prepare result to the user. This design relies on
readyset-mysql and readyset-psql's behaviors of using the upstream
prepare result to respond to a prepared statement request, when both an
upstream result and readyset result rae available.

The check for whether we can return a NoSchema result is made in the
adapter after we have attempted to get an upstream result. This is done
so that we return an error if the preparing backend has temporarily lost
connection to upstream.

Change-Id: I10c9747e687be62b3625af1e84a7d0ce11b448d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3846
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>