dataflow: Clarify naming + add docstrings for InitialState
The naming in the InitialState enum has long been pretty diverged from
the naming we use both elsewhere in the code and verbally to talk about
nodes and their state - among other things, this enum referred to a
reader node as "global" and an internal node as "local", which I'm not
entirely clear on the original reasoning for. Regardless, it's best to
have consistent naming for things, so this commit renames a lot of the
variants and fields of this enum for improved clarity, and adds
docstrings while we're at it.

Change-Id: I68cf39759c1856805e1cf64c905dff3fc1086215
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3821
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>