Rename readyset-client to readyset-adapter
We'd like to combine the `readyset-psql` and `readyset-mysql` binaries
into a single binary called `readyset`. We've already got a crate named
`readyset`, which contains client bindings for the readyset server - I'd
like to call that `readyset-client` to get the name out of the way, but
we've already got a crate called `readyset-client` which contains shared
code for adapters! To get *that* name out of the way, then, this commit
renames `readyset-client` to `readyset-adapter`. This creates
some (temporary) confusion with the `readyset-client-adapter` crate -
when all this is finished my plan is to merge that code into the
eventual `readyset` combined crate once that exists, since it contains
the command-line options for what that binary will eventually contain.

Refs: ENG-2137
Change-Id: I1e8bbaba1a672d03c0a840813651bc92cd637e75
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3831
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Tested-by: Buildkite CI