Rename readyset to readyset-client
We're renaming a bunch of crates with the eventual plan of merging
together `readyset-mysql` and `readyset-psql` into a single binary
called `readyset`. To get out of the way of that name, this commit
renames the *existing* `readyset` crate to `readyset-client` (now that
the old `readyset-client` crate has been renamed to `readyset-adapter`).

Refs: ENG-2137
Change-Id: I855bb9c02c18ba2295d119fec03027f009545d55
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3833
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>