replicators: Fix "maximum offset must be present" panic
In situations where a table is present in ReadySet but no longer exists
upstream, we can run into problems where the table is not snapshotted
and thus no replication offset is set, yet we still try to get a
replication offset later while we're calculating the max offset, and the
table return None which ultimately triggers the aforementioned panic.

To fix this, this commit adds code to check for such tables and delete
them prior to beginning replication.

Change-Id: Ib5e080447f267125601d959db398e924db6657c2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3835
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>