nom-sql: Make TableExpr capable of being a subquery
Split out the `table` field for TableExpr into a new `TableExprInner`
enum, which can be either a table reference or a subquery (in the
future, we'll also add table *expressions*). Since this is redundant
with JoinRightSide::NestedSelect, this also removes that. The effect
here is that we can now parse subqueries *anywhere* we can parse table
expressions, including (now) in the FROM list of a SELECT statement in
addition to just the right hand side of joins. This doesn't
properly *support* SELECT FROM subquery yet, but paves the way towards
that.

Since this is a relatively foundational part of our AST, lots of stuff
had to change to support this - primarily lots of this is just pattern
matching on TableExprInner to figure out what kind of table expr we're
talking about - most of the interesting stuff here is in query_graph,
where we're also laying the groundwork for full support for SELECT FROM
subquery by trying to do the correct things with subquery TableExprs

Refs: ENG-168
Change-Id: I3b7e0f10a498ecd84a2f709b902db2a176a274c4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3861
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>