sql-passes: Consider FROM clause for subquery schemas
Now that the FROM clause can contain subqueries, pass in a list of table
exprs in the FROM clause to `utils::subquery_schemas` and consider those
as part of the subqueries to return. This causes SELECT FROM subquery to
work with unqualified column references to those subqueries, and
`*`-expansion including columns from those subqueries (so this also
includes a logictest for that).

Refs: ENG-168
Change-Id: Ib35d91285f5e6eaaf4423599756afd9a8e28661c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3862
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>