readyset-data: Add `Array::outer_dimension` method
This method returns a new `ArrayView` type, which enables iterating over
rows of a 2D array. This type can be reused in the future for other
axes.

The `ArrayViewD::iter` method is limited to returning a lifetime over
itself. As a result, `&DfValue` cannot be used outside of the loop
`ArrayView` is created in. The initial version of this change extended
the lifetime to be over `&'v Array`/`&'v DfValue` with `&'a ArrayView`
where `'v: 'a`. However the team deemed this use of `unsafe` was
unwarranted when `DfValue::clone` is considered to be cheap.

Change-Id: I56ad4070b6a058fa87e1d27dc4aa83c23d2df12c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4022
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI