query_graph: Error on duplicate table names
When adding new relations to a query graph during construction, return
an error if the same table name (or alias) exists twice in the query, to
match both MySQL's and PostgreSQL's behavior. Even though we technically
don't *have* to do this (we aren't currently trying to reject all
invalid queries) it's relatively trivial to do, and keeps us a little
bit more durable against this kind of thing being accidentally triggered
by *us*, if not by a user.

Change-Id: I33df5574ae9dfa90340f4de114095a7df9b59582
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3873
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>