Inlining: Multi-view support for view_builder rpc
Updates the view_builder RPC to allow for the controller to return
mutliple reused views for a view request. This CL provides the
infrastructure but does not yet return multiple views in the rpc.

For example, if we have multiple migrations with the expression
'SUM(<col>) = <value>', with different 'value's, we can return a view
for each of these migrations when a client requests a view for the same
query, but with the expression 'SUM(<col>) = ?'.

To allow for this, the `View` struct was wrapped with a new `View` enum,
that either contains a `ReaderHandle` (which was formerly the `View`) or
a `Vec1<ReusedReaderHandle>`. The `required_values` and `key_remapping`
fields were removed from `View` and given to a new `ReusedReaderHandle`
struct so that a `ReaderHandle` no longer represents a reused cache by
itself. Similar changes were made to the `ViewBuilder` struct.

When executing a query against multiple reused caches, the `View` will
handle building a `ViewQuery` for each `BorrowedView` that it has,
stopping after the first success. This was accomplished by moving
`build_view_query()` into `View`.

Additionally, short_circuit_empty_resultset() was removed from the
adapter, as it was created for an impractical edge case and would have
required a refactor to maintain the new `View` abstraction.

Change-Id: Idd2ea00ab23ea811559dc3e405114408297b8760
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3884
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>