client-adapter: Stop using RPC to get pagination support
The existing code that does an RPC to ask the server about pagination
support can time out and panic if the server takes too long to respond.
This can easily become a problem in embedded mode, since the server and
adapter are started at the same time, and the controller often blocks on
server startup while it rebuilds the dataflow graph (which can
particularly take a while if there are fully materialized queries that
require a lot of data to be replayed).

To fix this, this commit has the client grab the relevant config flags
directly from the startup options. In this case the relevant flags are
set via env vars, which means that we now need to make sure to set the
same flag values on the client as well as the server, but since these
are undocumented experimental flags anyway, this doesn't seem like a
huge drawback (and it also doesn't matter in the case of standalone
mode, of course).

Change-Id: I604f449d869461c4aadad87f7074f1784bf809ea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3886
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>