Ensure StandaloneAuthority respects DB_DIR
Currently StandaloneAuthority always stores rocksdb database files in
the current directory. It should instead respect whatever is set for
DB_DIR, so the files are always stored in the same location as the
readyset state files for easier cleanup.

This commit updates behavior so that if DB_DIR is passed in, we will use
that instead of the default of the current directory.

Change-Id: I2abfd387e9ad17657990edd44795b853ff5a5835
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3898
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>