readyset-dataflow: Consider failed ops errors
This logs and returns an error if a base node enounteres a failed op
rather than warning and continuing.

If a base node hasn't fully applied every op that we don't have any
mechanism to reconcile our state, so we will be permanently diverged
from the correct state. Consider this an error and do not continue
processing.

In order to provide more detail in the error and logs, we bubble up the
failed ops to the call site by passing the FailedOpLogger in the error,
then inject the node name from the call site.

Also, Base::process was renamed to Base::process_ops since there are a
_lot_ of process functions and it was hard to find the relevant one
without just renaming it.

Release-Note-Core: Consider failing to apply operations to the base
  nodes an error.
Change-Id: I5c0811094769ce03ecc1ebd8db4cb970186a9caa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3897
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>