mir: Simplified and renamed `FlowNode`.
The `FlowNode` enum had two variants: `New` and `Existing`.
However, no matter which one was used, it was always changed right away
to `Existing`, so it made no sense to keep both and make it harder to
understand what it was used for.
This commit simplifies `FlowNode` to be a struct holding only the
`NodeIndex`, and adds documentation.

This commit also renames `FlowNode` to be `DfNodeAddress`, so the naming
 is consistent with other Dataflow related structs (ie, `DfState`,
`DfType`, etc).

Refs: ENG-1634
Change-Id: I6426aed1403d5942c0956e7a6ceff01a719b6988
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3904
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>