recipe: Add Change to record non-replicated tables
Add a new recipe Change to insert a Relation into a set of tables that
are known to exist in the upstream, but are not being replicated (either
because the user opted out of replicating them, or because something
about replicating or snapshotting them failed for some reason). Also,
update the Drop change to properly drop tables from this set before
looking in custom types or other relations we know about.

As of this commit, nothing actually *sends* this change - that's coming
next.

Refs: ENG-2134
Change-Id: Ifd9d83f966d87340b842008bdbe35410b54c9452
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3913
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>