postgres replicator: Support TOAST in unkeyed tables
This change adds support for replicating Postgres WAL update events
containing "unchanged" TOAST values, but only on unkeyed tables with
REPLICA IDENTITY FULL. It does so by overwriting the unchanged value
(which we still don't support) in the new row with the corresponding
value from the old row (which the WAL supplies in full, because there's
no key).

This leaves intact the logic from the previous change to skip
TOAST-containing tables during snapshot. We can take that out once we
support TOAST in all cases.

Change-Id: I8db941e896cd2f08aebfaad14177927bd77ec4a7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3908
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>