readyset-tracing: Add configurable ctx to logs
This wraps the tracing log macros with our own and allows for a way to
inject arbitrary context (for example the deployment id, to be added in
a follow-up change) into each log as a field.

Since this field is set at runtime, this introduces a way to set
it before logging begins.

Refs: ENG-2102
Change-Id: I0649931f4e21bd2f5349df8b89b1d73c54a4ce50
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3915
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>