server: Remove ALTER TABLE handling from recipe
Currently, all ALTER TABLE changes require resnapshotting, meaning they
won't actually be sent as part of a Changelist, and instead we'll just
get the CREATE TABLE change for the table post-alter. This means that
the code to handle AlterTable in recipe is unnecessary. This commit
removes it, replacing it with an internal error return, so we don't have
to pay the cost of maintaining it.

Change-Id: I30e58704e9ad9d1c4ada48b39965ef8aef6c5848
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3918
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>