nom-sql: Make CREATE TABLE parsing fallible
Similar to what we're doing for the ALTER TABLE statement, this commit
makes the parsing of CREATE TABLE statements fallible, meaning we can
parse the `CREATE TABLE IF NOT EXISTS <name>` bit, then fail on either
the body of the statement (with fields and keys), the options, or some
combination of the two. Similar to AlterTableStatement, this involves
changing the AST to contain a Result where the Ok case contains the
successfully parsed AST, and the Err case contains the full string that
failed to parse. To make this work, this breaks out a new
CreateTableBody struct with `fields` and `keys`. Outside of nom-sql,
most of this change actually consists of changing places where we're
storing a full CreateTableStatement to instead store that
CreateTableBody, since that actually contains the data we need and it
avoids having to do full error handling every time we use it. This moves
the needle at least *a little bit* on an overarching goal I have to use
the AST less pervasively throughout internal structures.

As of this commit, we don't actually *do* anything with this fallibility
- if a CreateTableStatement containing an Err for options or body is
passed to the `extend_recipe` RPC we just return an error - but next up
we'll be using this along with the new AddNonReplicatedTable change to
record that these tables exist in the upstream even if we don't have
base nodes for them, so that they correctly participate in schema
resolution.

Fixes: ENG-2169
Change-Id: I518055e448e098c946e532b19a2e601991304145
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3920
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>