mir: Added function to `MirGraph` to resolve Dataflow nodes.
Since we don't intend to assign dataflow nodes to all MIR nodes
(particularly, leaf and alias MIR nodes), we will need a way to resolve
the dataflow nodes associated with a given MIR node during the MIR to
Dataflow transformation.

Ideally, we shouldn't need to resolve those nodes ourselves, but rather
keep track of the last dataflow nodes we added during that process (see
ENG-2175).
Since that would require a bigger refactor, for now we do the resolution
 ourselves.

Refs: ENG-1634
Change-Id: I7e88ff4e7fac51b7d6a7bb46eeba58cef6c49db6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3925
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>