dataflow-state: Make State::process_records return a Result
State::process_records is *actually* fallible in the case of base tables
- for example, we can receive delete events for records we don't know
about if something's gone *very* wrong during replication (and this has
happened in production, for reasons we're still trying to figure out).
Currently, if anything fails, we just panic, which ends up bubbling up
to the "abort on panic" handling which aborts the process if *any*
domain thread panics. We can do better than this in this case - if we do
get a bad batch of deltas for a table, we can react to this in the
replicators by resnapshotting the table from scratch.

This commit starts paving the road towards making that possible,
starting out by just making State::process_records able to return a
Result, and threading that through in various necessary
places (including making Packet::map_data more idiomatically just a
`mut_data` accessor function that returns a mutable reference...).
Currently there are no code paths that return errors here, but that's
coming up soon.

Change-Id: I36658a2f22f12e25c211d0bbf71db0ea26c0b1b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3928
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>