sql-passes: Use non-replicated-relations in schema resolution
If, during schema resolution, an unqualified table reference resolves to
a relation that is known to exist in the upstream db but is not being
replicated, return an error rather than continuing to resolve schemas.
This is important for correctness because if this table *was* being
replicated, the query would select from it instead of a table with the
same name later in the search path.

Refs: ENG-2135
Change-Id: Ib01859dd1cf20ba05f5a901e6253eca4ac4597d3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3934
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>