telemetry: Add quickstart DEPLOYMENT_ENV variants
This change adds the variants "quickstart_docker" and
"quickstart_instruqt" to the DEPLOYMENT_ENV env var, so we can correlate
telemetry with whether the user followed one of our guides.

It also leverages serde_json to parse the env var. DEPLOYMENT_ENV
variants will always be expected as snake_case.

Change-Id: I1e1fd1cfd2094e315ba37a51a0f170e106193fb6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3944
Tested-by: Buildkite CI
Reviewed-by: Sif Hall <sif@readyset.io>