Rename /prometheus to /metrics
The Prometheus/Otel ecosystem defaults to /metrics for the scrape path.
This change updates the readyset internal path from /prometheus to
/metrics and updates the dependencies to match.

This is worthy of a release note to the public as it enables people to
minimize their configuration with the removal of the metrics_path option
but also requires them TO make that update.

Release-Note-Core: Change metrics exposition endpoint from /prometheus
  to /metrics for better alignment with the Prometheus/Otel ecosystem.
  Operators will need to update their Prometheus configurations to
  change the option `metrics_path: "/prometheus"` to `/metrics`, or
  remove the option entirely as the default for Prometheus *is*
  `/metrics`. :smile:

Change-Id: Ic1c140ec65eb9ef46a33c977c3d4de0b85b60b6d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3947
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>