replicators: Move MAX_QUEUED_ACTIONS check
This moves the check to see whether we've exceeded MAX_QUEUED_ACTIONS to
after we've attempted to get the next event, and if we've exceeded the
limit we store the next event in self.peek to be picked up next time
around.

This shouldn't result in any real change in functionality, but it lines
us up for the next change I'm working on, which requires us to be able
to peek at the next incoming event before we decide whether to return
the current batch of actions or not.

Change-Id: Idd0be54399cdc0dcc6a1e7987d95339ac9503f66
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3953
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>