replicators: Fix batching for WAL events that share the same LSN
It's possible for multiple consecutive events to arrive from Postgres
that share the same LSN; currently the one case we know of that can
cause this is use of the COPY command (and \copy by association, since
it uses COPY under the hood). COPY uses the Postgres `heap_multi_insert`
function under the hood, which appears to combine multiple inserts into
a single WAL operation for efficiency.

Prior to this fix, we could *sometimes* end up incorrectly skipping
events because we use the LSN of the final event in a batch to determine
whether we've already seen that batch of events or not, and this code
was written under the assumption that future events would always have
higher LSNs than past ones; if we got a long series of events that
shared the same LSN though, then it was possible for the final event in
two consecutive batches to share the same LSN, which would cause the
second batch to be skipped, resulting in dropped inserts.

To fix this, we never end a batch unless the next event has a different
LSN. This should generally be fine, since we can always expect an event
with a new LSN to arrive right away after a series of same-LSN events,
as we'll either receive more inserts with new LSNs or a Commit event
with a new LSN if the COPY is finished. It can potentially result in
pretty long batches of events, since I've seen over 130 events come in
in a row with the same LSN, and if a ~130-long batch starts near the end
of the 100-event cutoff, we could get batches longer than 230 events.
But this should be relatively rare in practice, and if this causes any
performance issues, we could improve this code to split up batches more
intelligently along LSN-change boundaries.

This commit also contains a replication test which reliably fails
without the fix contained in this commit, but which passes with this fix
applied.

Release-Note-Core: Fixed a bug that could lead to missing cache data
  and/or crashes after inserting large amounts of table data with the
  Postgres COPY or \copy commands.
Fixes: ENG-2100
Change-Id: I9449f680b33d693ec3c22085314c3a847a677445
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3958
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>