nom-sql: Make CREATE VIEW parsing fallible
Just the same as we did for CREATE TABLE and ALTER TABLE, this commit
makes the parsing for the CREATE VIEW statement fallible, allowing us to
parse everything up to the name of the view without successfully parsing
the actual definition. This is being done for the same reason as CREATE
TABLE - so that we can record views that fail to parse within the set of
non-replicated relations, so that they correctly participate in schema
resolution for unqualified relation references in queries.

Fixes: ENG-2192
Change-Id: I7434e1a2f96c4763835609f501579b52030a10fd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3960
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>