treewide: Fix all clippy lints in all targets, and run in CI
Clippy has the ability to catch lots of issues that are very relevant to
tests and benchmarks, such as failing to call `unwrap()` on results, or
using `assert(x.is_ok())` instead of `x.unwrap()` (the latter has much
better error messages when it fails). Since we've never run clippy
against anything other than lib or bin code, there's a lot of clippy
lints that have to be fixed before we start running clippy on tests and
benchmarks in CI - this commit fixes all of them, and then adds
`--all-targets` to the clippy command that runs in CI to run clippy
against tests (we also have to add `--features bench` since there's a
bench target that depends on that feature flag). Since nothing was even
checking that the benchmarks *compiled*, some things had to update in
some benchmarks to get them compiling after breaking changes that we
missed.

Refs: ENG-2134
Change-Id: I8e62807076583a7129f2347878806a14694d982f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/3961
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Luke Osborne <luke.o@readyset.io>